fix: GcpFirestoreEmulatorAutoConfiguration constructs invalid document parent
Fixes: #2290.
Fixes: #2330.
